This has become a utility class.
